requirements.txt
appnope==0.1.2
backcall==0.2.0
cffi==1.14.5
decorator==5.0.6
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
msgpack==1.0.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
py-algorand-sdk==1.4.1
pycparser==2.20
pycryptodomex==3.10.1
Pygments==2.8.1
PyNaCl==1.4.0
pyteal==0.7.0
six==1.15.0
traitlets==5.0.5
wcwidth==0.2.5
